requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
click==6.7
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
Flask==0.12.2
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
lxml==3.8.0
Markdown==2.2.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
numpy==1.13.0
olefile==0.44
opencv-python==3.3.0.9
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.3.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.10
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.2
